requirements.txt
speedtest-cli>=2.1.3
rich>=13.0.0
